requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
plotly==5.17.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio
datetime
json
time
urllib3==2.0.7
